Address review comments from @IngridPuppet
