+++ src/main/java/io/github/adorsysgis/keycloak/protocol/oid4vc/oid4vp/authenticator/mdoc/MdocCredentialVerifier.java
@@ -81,6 +81,10 @@ public class MdocCredentialVerifier implements CredentialVerifier {
 
         if (authReqs.shouldEnforceRevocationStatus()) {
             try {
+                // Status is stored in the MSO payload per IETF Token Status List §Referenced Token
+                // (https://www.ietf.org/archive/id/draft-ietf-oauth-status-list-11.html#name-referenced-token-in-cose)
+                // and MATTR docs
+                // (https://learn.mattr.global/docs/holding/credential-claiming-guides/revocation-status-check).
                 tokenStatusValidator.validate(verificationContext.getVerifiedMsoPayload());
             } catch (ReferencedTokenValidationException e) {
                 throw new VerificationException(
